Optimizations
Integration Test: Init

Integration tests: Preparation
