--- generate_arb.bat
@@ -1,1 +1,0 @@
-flutter packages pub run intl_translation:extract_to_arb --output-dir=assets/l10n lib/localizations.dart
