--- generate_arb.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env sh
-flutter packages pub run intl_translation:extract_to_arb --output-dir=assets/l10n lib/localizations.dart
