+++ lib/core/constants/app_constants.dart
@@ -3,7 +3,7 @@ class RoutePaths {
   static const String Authentication = '/';
   static const String Reporting = '/reporting';
   static const String Docs = '/docs';
-  static const String Calendar = '/calendar';
+  static const String Agenda = '/agenda';
   static const String About = '/about';
   static const String Feedback = 'feedback';
   static const String News = '/news';
