+++ lib/ui/views/portal_tabs/portal_apps/sogo.dart
@@ -6,13 +6,6 @@ class SogoWebView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // return WebviewScaffold(
-    //   appBar: AppBar(
-    // backgroundColor: Colors.green,
-    // title: const Text("Sogo"),
-    //   ),
-    //   url: 'https://sogo.emse.fr/',
-    // );
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Colors.green,
