+++ lib/ui/widgets/base_view_widget.dart
