+++ test_integration.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+flutter drive --target=test_driver/app.dart
